Create answers tables for questions

+++ database/migrations/2023_11_17_144515_create_answers_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('question_id');
+            $table->unsignedBigInteger('question_id');
             $table->unsignedInteger('user_id');
             $table->text('body');
             $table->integer('votes_count')->default(0);
